Misc cleanup around the area
