+++ oreClient/package.json
@@ -10,11 +10,14 @@
     "@fortawesome/free-regular-svg-icons": "^5.13.0",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "@fortawesome/vue-fontawesome": "^0.1.9",
+    "clipboard": "^2.0.6",
+    "filesize": "^6.1.0",
     "lodash": "^4.17.15",
     "markdown-it": "^10.0.0",
     "markdown-it-anchor": "^5.2.7",
     "markdown-it-task-lists": "^2.1.1",
     "markdown-it-wikilinks": "^1.0.1",
+    "moment": "^2.24.0",
     "nprogress": "^0.2.0",
     "query-string": "^6.12.1",
     "vue": "^2.6.11",
