+++ oreClient/src/main/assets/utils.js
@@ -68,3 +68,9 @@ export function avatarUrl(name) {
         return formatString(config.security.api.avatarUrl, name)
     }
 }
+
+export function numberWithCommas(x) {
+    const parts = x.toString().split(".");
+    parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+    return parts.join(".");
+}
