+++ orePlayCommon/app/views/layout/base.scala.html
@@ -15,18 +15,14 @@ this.
     <head>
         <link href="@config.security.api.url" rel="preconnect" crossorigin>
         <link href="@config.app.discourseCdnUrl" rel="preconnect" crossorigin>
-        <link href="@config.security.api.url" rel="dns-prefetch" >
-        <link href="@config.app.discourseCdnUrl" rel="dns-prefetch" >
 
-        <link rel="prefetch" href="@assetsFinder.path("images/favicon.ico")">
-        <link rel="prefetch" href="https://fonts.googleapis.com/css?family=Montserrat:400,700">
-        <link rel="prefetch" href="@assetsFinder.path("highlight/styles/default.css")">
-        <link rel="prefetch" href="@assetsFinder.path("bootstrap/css/bootstrap.min.css")">
-        <link rel="prefetch" href="@assetsFinder.path("lib/ore-client/commons.css")">
-        <link rel="prefetch" href="@routes.Assets.versioned("lib/ore-client/main.css")">
-        <link rel="prefetch" href="@assetsFinder.path("lib/ore-client/font-awesome.css")">
-        <link rel="prefetch" href="@assetsFinder.path("manifest/manifest.json")">
-        <link rel="prefetch" href="@routes.Assets.versioned("lib/jquery/dist/jquery.min.js")">
+        <link rel="preload" href="https://fonts.googleapis.com/css?family=Montserrat:400,700" as="style">
+        <link rel="preload" href="@assetsFinder.path("highlight/styles/default.css")" as="style">
+        <link rel="preload" href="@assetsFinder.path("bootstrap/css/bootstrap.min.css")" as="style">
+        <link rel="preload" href="@assetsFinder.path("lib/ore-client/commons.css")" as="style">
+        <link rel="preload" href="@assetsFinder.path("lib/ore-client/main.css")" as="style">
+        <link rel="preload" href="@assetsFinder.path("lib/jquery/dist/jquery.min.js")" as="script">
+        <link rel="preload" type="text/javascript" href="@routes.Application.javascriptRoutes" as="script">
 
         <title>@title</title>
 
@@ -41,14 +37,14 @@ this.
         <link rel="stylesheet" type="text/css" href="@assetsFinder.path("bootstrap/css/bootstrap.min.css")" />
         <link rel="stylesheet" type="text/css" href="@assetsFinder.path("lib/ore-client/vendors.css")" />
         <link rel="stylesheet" type="text/css" href="@assetsFinder.path("lib/ore-client/commons.css")" />
-        <link rel="stylesheet" type="text/css" href="@routes.Assets.versioned("lib/ore-client/main.css")" />
+        <link rel="stylesheet" type="text/css" href="@assetsFinder.path("lib/ore-client/main.css")" />
         @additionalStyling
 
         <!-- See https://goo.gl/OOhYW5 and https://goo.gl/qRE0vM -->
-        <link rel="manifest" href="@assetsFinder.path("manifest/manifest.json")">
+        <link rel="manifest" href="@assetsFinder.path("manifest/manifest.webmanifest")">
         <meta name="theme-color" content="#333">
 
-        <script type="text/javascript" src="@routes.Assets.versioned("lib/jquery/dist/jquery.min.js")"></script>
+        <script type="text/javascript" src="@assetsFinder.path("lib/jquery/dist/jquery.min.js")"></script>
     </head>
     <body @if(!empty) {class="site"}>
         @if(empty) {
@@ -101,10 +97,7 @@ this.
             <script type="text/javascript" src="@assetsFinder.path("lib/ore-client/vendors.js")"></script>
             <script type="text/javascript" src="@assetsFinder.path("lib/ore-client/font-awesome.js")"></script>
             <script type="text/javascript" src="@assetsFinder.path("lib/ore-client/commons.js")"></script>
-            <script type="text/javascript" src="@assetsFinder.path("lib/moment/min/moment.min.js")"></script>
             <script type="text/javascript" src="@assetsFinder.path("bootstrap/js/bootstrap.min.js")"></script>
-            <script type="text/javascript" src="@assetsFinder.path("lib/filesize/lib/filesize.js")"></script>
-            <script type="text/javascript" src="@assetsFinder.path("lib/clipboard/dist/clipboard.min.js")"></script>
             <script type="text/javascript" src="@assetsFinder.path("highlight/highlight.pack.js")"></script>
             <script type="text/javascript" src="@assetsFinder.path("javascripts/main.js")"></script>
             <script type="text/javascript" src="@assetsFinder.path("javascripts/apiRequests.js")"></script>
