+++ project/WebjarsDeps.scala
@@ -3,11 +3,8 @@ import sbt._
 //noinspection TypeAnnotation
 object WebjarsDeps {
 
-  val jQuery      = "org.webjars.npm" % "jquery"       % "2.2.4"
-  val fontAwesome = "org.webjars"     % "font-awesome" % "5.12.0"
-  val filesize    = "org.webjars.npm" % "filesize"     % "6.0.1"
-  val moment      = "org.webjars.npm" % "moment"       % "2.24.0"
-  val clipboard   = "org.webjars.npm" % "clipboard"    % "2.0.6"
-  val chartJs     = "org.webjars.npm" % "chart.js"     % "2.9.3"
-  val swaggerUI   = "org.webjars"     % "swagger-ui"   % "3.25.0"
+  val jQuery    = "org.webjars.npm" % "jquery"     % "2.2.4"
+  val moment    = "org.webjars.npm" % "moment"     % "2.24.0"
+  val chartJs   = "org.webjars.npm" % "chart.js"   % "2.9.3"
+  val swaggerUI = "org.webjars"     % "swagger-ui" % "3.25.0"
 }
